requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.2
aiosignal==1.3.1
aiosqlite==0.20.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bitarray==2.9.3
bitstring==4.1.4
bleach==6.2.0
boto3==1.34.46
botocore==1.34.162
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
chia-blockchain==2.3.0
chia_dev_tools==1.2.6
chia_rs==0.6.1
chiabip158==1.4
chiapos==2.0.4
chiavdf==1.1.4
click==8.1.3
clvm==0.9.9
clvm_tools==0.4.9
clvm_tools_rs==0.1.40
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
concurrent-log-handler==0.9.25
contourpy==1.3.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1
django-browser-reload==1.13.0
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
dnslib==0.9.24
dnspython==2.5.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_git==0.51.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==24.3.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numpy==2.2.4
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==3.3.0
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.3.0
redis==5.2.1
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
s3transfer==0.10.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setproctitle==1.3.3
setuptools==75.5.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.6.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.32.1
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.8.2
yarl==1.17.1
zipp==3.21.0
zstd==1.5.5.1
